requirements.txt
fastapi==0.115.12
fitz==0.0.1.dev2
nltk==3.9.1
numpy==2.2.6
pandas==2.0.3
Pillow==10.0.0
Pillow==11.2.1
pydantic==1.10.11
PyMuPDF==1.22.5
pytesseract==0.3.10
pytesseract==0.3.13
scikit_learn==1.4.0
sentence_transformers==3.4.1
textstat==0.7.5
uvicorn==0.34.2
